+++ src/constant/uiKit/index.ts
@@ -1,2 +1,4 @@
 /** SVG 图标统一的宽度 */
 export const svgStockeWidth = 1.5;
+/** 默认消息提示点背景色 */
+export const badgeOverlayColor = 'orange';